data/monthly_scheduler_trigger/requirements.txt
google-cloud-pubsub==2.20.0
Flask==2.3.2